PQRS - Cargar Excel para registrar las PQRS

+++ Modules/PQRS/Database/Migrations/2024_05_16_113306_create_pqrs_table.php
@@ -16,15 +16,15 @@ class CreatePqrsTable extends Migration
         Schema::create('pqrs', function (Blueprint $table) {
             $table->id();
             $table->foreignId('type_pqrs_id')->constrained()->onDelete('cascade');
-            $table->integer('filing_number');
+            $table->unsignedBigInteger('filing_number');
             $table->date('filing_date');
-            $table->unsignedInteger('nis');
+            $table->unsignedBigInteger('nis');
             $table->date('start_date');
             $table->date('end_date');
             $table->text('issue');
             $table->enum('state', ['ARCHIVADO', 'EN PROCESO', 'PROXIMO A VENCER', 'RESPUESTA GENERADA', 'RESPUESTA PARCIAL'])->default('EN PROCESO');
             $table->text('answer')->nullable();
-            $table->integer('filed_response')->nullable();
+            $table->bigInteger('filed_response')->nullable();
             $table->date('response_date')->nullable();
             $table->softDeletes();
             $table->timestamps();
